fix: 修复云端烤群友 CD 时间戳精度异常
- user_usage.last_roast_ts 改为 BIGINT，避免 MySQL FLOAT 存储 Unix 时间戳时产生精度误差
- 云端冷却计算统一按整数秒处理，修复普通烤群友 8 小时 CD 被算成 9 小时以上的问题
- 旧数据导入脚本同步改为写入整数秒时间戳，避免再次引入相同问题

+++ rollpig_cloud/models.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import datetime as dt
 
-from sqlalchemy import Date, DateTime, Float, Index, String, UniqueConstraint, func
+from sqlalchemy import BIGINT, Date, DateTime, Index, String, UniqueConstraint, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from .db import Base
@@ -52,7 +52,7 @@ class UserUsage(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     user_id: Mapped[str] = mapped_column(String(64), nullable=False, unique=True, index=True)
-    last_roast_ts: Mapped[float | None] = mapped_column(Float, nullable=True)
+    last_roast_ts: Mapped[int | None] = mapped_column(BIGINT, nullable=True)
     last_force_date: Mapped[dt.date | None] = mapped_column(Date, nullable=True)
     updated_at: Mapped[dt.datetime] = mapped_column(DateTime, nullable=False, server_default=func.now(), onupdate=func.now())
